refactor test to avoid multiple execution of inherited tests
